Navigation fixed and switched to MVVM
The navigation graphs have been splitted into the main navigation graph
and the home navigation graph, so that the NavHostFragment can be used
on the home component.

The MVVM architecture has been put in place replacing the MVP
architecture that was prevised. The MVVM architecture is a common and
mainstream practice on android development.